fixed warning-as-error on windows for OPAL_SHORT_ATOM

+++ Common/DtaCommand.cpp
@@ -177,7 +177,7 @@ void
 DtaCommand::addToken(OPAL_UID token, OPAL_UID_SIZE size)
 {
     LOG(D1) << "Entering DtaCommand::addToken(OPAL_UID)";
-    cmdbuf[bufferpos++] = size==OPAL_UID_SIZE::HALF_UID?OPAL_SHORT_ATOM::BYTESTRING4:OPAL_SHORT_ATOM::BYTESTRING8;
+    cmdbuf[bufferpos++] =(uint8_t)( size==OPAL_UID_SIZE::HALF_UID?OPAL_SHORT_ATOM::BYTESTRING4:OPAL_SHORT_ATOM::BYTESTRING8 );
     memcpy(&cmdbuf[bufferpos], &OPALUID[token][0], size==OPAL_UID_SIZE::HALF_UID?4:8);
     bufferpos += size==OPAL_UID_SIZE::HALF_UID?4:8;
 }
